Move standards skills relationships to separate controller #5
